==> notebooks/1.1-jalvaradoruiz-extraccion-data-pjud.ipynb <==
# Extraccion datos 

Desde https://numeros.pjud.cl/Descargas

Extraccion de los datos referentes a causas penales desde 2015 a 2019, en ella se encuentran datos separados por:

- Ingresos y terminos por rol y materia
- Inventarios de Causas
- Duración de Causas
- Audiencias de cada Causa

REV: 16-10-2020 

import requests
from multiprocessing.pool import ThreadPool


materias = ['penal']
categorias = ['Ingresos','Terminos','Inventario','Duracion','Audiencias']


años = list(range(2015,2020))
files = []


def url_response(url):
    path, url = url
    descarga = requests.get(url, stream = True)
    with open(path, 'wb') as f:
        for chunk in descarga:
            f.write(chunk)

for año in años:
    for materia in materias:
        for categoria in categorias:           
            url = f"https://numeros.pjud.cl/descargas/{materia}/{categoria}/{categoria}-{año}-CSV.zip"
            name_file = f"{materia}-{categoria}-{año}.zip"
            files.append((name_file,url))
     
    
# Descargo archivos           
ThreadPool(9).imap_unordered(url_response, files)



# Mover los archivos zip a Carpeta Download (se dejan afuera los archivos zip vacios)
import os
import shutil

os.mkdir("../data/external/zip")

archivos = os.listdir(".")

for archivo in archivos:
    path = os.getcwd() + "/" + archivo
    tamaño_archivo = os.path.getsize(path)
    if archivo.find(".zip") != -1 and tamaño_archivo != 2487:
        shutil.move(archivo,'../data/external/zip/'+archivo)
