==> notebooks/2.1-jalvaradoruiz-descompresion-data-pjud.ipynb <==
# Descomprimir archivos Zip

Extraemos la Data y la guardamos en Folder raw


import os
import zipfile

path_extraccion = "../data/raw/pjud"
path_zip = "../data/external/zip"
archivos = os.listdir(path_zip) 

for archivo in archivos:
    path_archivo_zip = path_zip + "/" + archivo
    archivo_zip = zipfile.ZipFile(path_archivo_zip,"r")
    archivo_zip.extractall(pwd=None, path=path_extraccion)
    archivo_zip.close()

# Elimino carpeta Download con archivos ZIP
import shutil
shutil.rmtree(path_zip, ignore_errors=True)
