==> notebooks/3.1-jalvaradoruiz-carga-limpieza-data-ingresos-materia.ipynb <==
# Carga Data INGRESO X MATERIA

Creación del DataFrame asociado a los Ingresos por materia

Rev: 29-10-2020

import os
import pandas as pd
import numpy as np
from pyarrow import feather
from tqdm import tqdm

from pjud.data import cleandata

path_raw = "../data/raw/pjud" 
archivos = os.listdir(path_raw) 
tqdm.pandas()

#creacion dataframe con los datos de INGRESOS por MATERIA Penal de los años 2015 a 2019

dataframes = []

for archivo in archivos:
    if archivo.find("Ingresos por Materia Penal") != -1:
        df = pd.read_csv(f"{path_raw}/{archivo}", sep = ";", encoding = 'cp850', dtype = 'unicode',low_memory = True)
        dataframes.append(df) 
        
df_ingresos_materia = pd.concat(dataframes, axis = 0)


df_ingresos_materia

df_ingresos_materia.columns

Podemos observar que existen columnas que se repiten, y que tienen datos NAN en algunas pero esos datos en otras columnas, pasa en TOTAL INGRESOS POR MATERIAS=TOTAL INGRESOS POR MATERIAS(*)

df_ingresos_materia['TOTAL INGRESOS POR MATERIAS'] = df_ingresos_materia['TOTAL INGRESOS POR MATERIAS'].fillna(df_ingresos_materia['TOTAL INGRESOS POR MATERIAS(*)'])

df_ingresos_materia.drop(['N°','TOTAL INGRESOS POR MATERIAS(*)'], axis = 'columns', inplace = True)

df_ingresos_materia.drop(['(*)Se agregó columna total de ingresos, dado que en algunas causas, la materia se repite (error de tramitación)'], 
                         axis='columns',inplace=True)


# TRANSFORMAMOS DE FLOAT A INTEGER

df_ingresos_materia['COD. CORTE'] = df_ingresos_materia['COD. CORTE'].fillna(0).astype(np.int16)
df_ingresos_materia['COD. TRIBUNAL'] = df_ingresos_materia['COD. TRIBUNAL'].fillna(0).astype(np.int16)
df_ingresos_materia['COD. MATERIA'] = df_ingresos_materia['COD. MATERIA'].fillna(0).astype(np.int16)
df_ingresos_materia['AÑO INGRESO'] = df_ingresos_materia['AÑO INGRESO'].fillna(0).astype(np.int16)
df_ingresos_materia['TOTAL INGRESOS POR MATERIAS'] = df_ingresos_materia['TOTAL INGRESOS POR MATERIAS'].fillna(0).astype(np.int8)

# Transformamos fechas

df_ingresos_materia['FECHA INGRESO'] = df_ingresos_materia['FECHA INGRESO'].progress_apply(cleandata.convierte_fecha)

# Elimino espacios en las columnas tipo objetos

df_ingresos_materia = df_ingresos_materia.progress_apply(cleandata.elimina_espacios, axis=0)

# Elimino tildes 

cols = df_ingresos_materia.select_dtypes(include = ["object"]).columns
df_ingresos_materia[cols] = df_ingresos_materia[cols].progress_apply(cleandata.elimina_tilde) 

# Categorizacion

df_ingresos_materia['CORTE'] = df_ingresos_materia['CORTE'].astype('category')

df_ingresos_materia['TIPO CAUSA'].unique()

Dejo solo causas Ordinarias

tipo_causa = df_ingresos_materia[df_ingresos_materia['TIPO CAUSA']!='Ordinaria']

df_ingresos_materia.drop(tipo_causa.index, axis=0, inplace=True)

df_ingresos_materia.reset_index(inplace = True)
df_ingresos_materia

# Guardamos dataset como archivo feather
path_interim = "../data/interim/pjud"
os.makedirs(path_interim, exist_ok = True) 

df_ingresos_materia.to_feather(f'{path_interim}/clean_IngresosMateria.feather')
