==> notebooks/3.2-jalvaradoruiz-carga-limpieza-data-terminos-materia.ipynb <==
# Carga Data TERMINO X MATERIA

Creación del DataFrame asociado a los terminos por materia

Rev: 29-10-2020

import os
import pandas as pd
import numpy as np
from pyarrow import feather
from tqdm import tqdm

from pjud.data import cleandata

path_raw = "../data/raw/pjud" 
archivos = os.listdir(path_raw) 
tqdm.pandas()

#creacion dataframe con los datos de Términos por Materia Penal de los años 2015 a 2020

dataframes = []

for archivo in archivos:
    if archivo.find("Términos por Materia Penal") != -1:
        df = pd.read_csv(f"{path_raw}/{archivo}", sep = ";", encoding = 'cp850', low_memory = True)
        dataframes.append(df) 
        
df_termino_materia = pd.concat(dataframes, axis = 0)

df_termino_materia.columns

df_termino_materia['SISTEMA'].unique()

Elimino los registros asociados a Sistema METGE (METAS DE GESTION) ya que son causas para realizar cumplimiento de KPI

df_metge = df_termino_materia[df_termino_materia['SISTEMA']=='METGE']
df_termino_materia.drop(df_metge.index, axis=0, inplace=True)

# Estandarización de nombres de variables

df_termino_materia.rename(columns = {'CÓD. CORTE':'COD. CORTE',
                                     'CÓD. TRIBUNAL':'COD. TRIBUNAL',
                                     'CÓD. MATERIA':'COD. MATERIA',
                                     'MOTIVO DE TÉRMINO':'MOTIVO TERMINO',
                                     'DURACIÓN CAUSA':'DURACION CAUSA',
                                     'FECHA TÉRMINO':'FECHA TERMINO',
                                     'MES TÉRMINO':'MES TERMINO',
                                     'AÑO TÉRMINO':'AÑO TERMINO',
                                     'TOTAL TÉRMINOS':'TOTAL TERMINOS'
                                    },inplace = True)

df_termino_materia.drop(['N°','SISTEMA'], axis = 'columns', inplace = True)

# TRANSFORMAMOS DE FLOAT A INTEGER

df_termino_materia['COD. CORTE'] = df_termino_materia['COD. CORTE'].fillna(0).astype(np.int16)
df_termino_materia['COD. TRIBUNAL'] = df_termino_materia['COD. TRIBUNAL'].fillna(0).astype(np.int16)
df_termino_materia['COD. MATERIA'] = df_termino_materia['COD. MATERIA'].fillna(0).astype(np.int16)
df_termino_materia['DURACION CAUSA'] = df_termino_materia['DURACION CAUSA'].fillna(0).astype(np.int16)
df_termino_materia['AÑO TERMINO'] = df_termino_materia['AÑO TERMINO'].fillna(0).astype(np.int16)
df_termino_materia['TOTAL TERMINOS'] = df_termino_materia['TOTAL TERMINOS'].fillna(0).astype(np.int8)

# Transformamos formato fecha

df_termino_materia['FECHA INGRESO'] = df_termino_materia['FECHA INGRESO'].progress_apply(cleandata.convierte_fecha)
df_termino_materia['FECHA TERMINO'] = df_termino_materia['FECHA TERMINO'].progress_apply(cleandata.convierte_fecha)

# Elimino espacios en las columnas tipo objetos

df_termino_materia = df_termino_materia.progress_apply(cleandata.elimina_espacios, axis=0)

# Elimino tildes de object

cols = df_termino_materia.select_dtypes(include = ["object"]).columns
df_termino_materia[cols] = df_termino_materia[cols].progress_apply(cleandata.elimina_tilde)

df_termino_materia['RIT'] = df_termino_materia['RIT'].progress_apply(cleandata.limpia_rit)

# Categorizar variables

df_termino_materia['CORTE'] = df_termino_materia['CORTE'].astype('category')
df_termino_materia['MOTIVO TERMINO'] = df_termino_materia['MOTIVO TERMINO'].astype('category')

Dejo solo causas Ordinarias

df_termino_materia['TIPO CAUSA'].unique()

tipo_causa = df_termino_materia[df_termino_materia['TIPO CAUSA']!='Ordinaria']

df_termino_materia.drop(tipo_causa.index, axis=0, inplace=True)

# Reset el index para realizar feather

df_termino_materia.reset_index(inplace = True)

df_termino_materia

# Guardamos dataset como archivo feather
path_interim = "../data/interim/pjud"
os.makedirs(path_interim, exist_ok = True) 

df_termino_materia.to_feather(f'{path_interim}/clean_TerminosMateria.feather')
