==> notebooks/3.3-jalvaradoruiz-carga-limpieza-data-ingresos-rol.ipynb <==
# Carga Data INGRESOS X ROL

Creación del DataFrame asociado a los Ingresos por rol

Rev: 29-10-2020

import os
import pandas as pd
import numpy as np
from pyarrow import feather
from tqdm import tqdm

from pjud.data import cleandata

path_raw = "../data/raw/pjud" 
archivos = os.listdir(path_raw) 
tqdm.pandas()

# Creacion dataframe con los datos de Ingresos por Rol Penal de los años 2015 a 2019
dataframes = []

for archivo in archivos:
    if archivo.find("Ingresos por Rol Penal") != -1:
        df = pd.read_csv(f"{path_raw}/{archivo}", sep = ";", encoding = 'cp850', low_memory = True)
        dataframes.append(df) 
        print(f"Existen {len(df)} en archivo {archivo}")
        
df_ingresos_rol = pd.concat(dataframes, axis = 0)
#df_ingresos_rol.dtypes

# Transformamos variables float64 a int16

df_ingresos_rol['COD. CORTE'] = df_ingresos_rol['COD. CORTE'].fillna(0).astype(np.int16)
df_ingresos_rol['COD. TRIBUNAL'] = df_ingresos_rol['COD. TRIBUNAL'].fillna(0).astype(np.int16)
df_ingresos_rol['AÑO INGRESO'] = df_ingresos_rol['AÑO INGRESO'].fillna(0).astype(np.int16)
df_ingresos_rol.drop(['N°'], axis = 'columns', inplace = True)

df_ingresos_rol['FECHA INGRESO'] = df_ingresos_rol['FECHA INGRESO'].progress_apply(cleandata.convierte_fecha)

# Elimino espacios en las columnas tipo objetos

df_ingresos_rol = df_ingresos_rol.progress_apply(cleandata.elimina_espacios, axis=0)

# Elimino tildes de las columnas object

cols = df_ingresos_rol.select_dtypes(include = ["object"]).columns
df_ingresos_rol[cols] = df_ingresos_rol[cols].progress_apply(cleandata.elimina_tilde)

# Transformamos en variables categoricas

df_ingresos_rol['CORTE'] = df_ingresos_rol['CORTE'].astype('category')

Dejo solo causas Ordinarias

df_ingresos_rol['TIPO CAUSA'].unique()

tipo_causa = df_ingresos_rol[df_ingresos_rol['TIPO CAUSA']!='Ordinaria']

df_ingresos_rol.drop(tipo_causa.index, axis=0, inplace=True)

# Reset el index para realizar feather

df_ingresos_rol.reset_index(inplace = True)
df_ingresos_rol

# Guardamos dataset como archivo feather
path_interim = "../data/interim/pjud"
os.makedirs(path_interim, exist_ok = True) 

df_ingresos_rol.to_feather(f'{path_interim}/clean_IngresosRol.feather')
