==> notebooks/3.4-jalvaradoruiz-carga-limpieza-data-terminos-rol.ipynb <==
# Carga Data TERMINOS X ROL

Creación del DataFrame asociado a los Terminos por Rol

Rev: 29-10-2020

import os
import pandas as pd
import numpy as np
from pyarrow import feather
from tqdm import tqdm

from pjud.data import cleandata

path_raw = "../data/raw/pjud" 
archivos = os.listdir(path_raw) 
tqdm.pandas()

# Creacion dataframe con los datos de Términos por Rol Penal de los años 2015 a 2019

dataframes = []

for archivo in archivos:
    if archivo.find("Términos por Rol Penal") != -1:
        df = pd.read_csv(f"{path_raw}/{archivo}", sep = ";", encoding = 'cp850', low_memory = True)
        dataframes.append(df) 
        
df_termino_rol = pd.concat(dataframes, axis = 0)

df_termino_rol.columns

df_termino_rol['SISTEMA'].unique()

Elimino causas que no sean SIAGJ

df_no_siagj = df_termino_rol[df_termino_rol['SISTEMA']!='SIAGJ']

df_termino_rol.drop(df_no_siagj.index, axis=0, inplace=True)

# Elimino filas vacias o con datos NaN

df_termino_rol = df_termino_rol.dropna()
df_termino_rol.drop(['N°','SISTEMA'], axis = 'columns', inplace = True)

# Cambio de nombre a algunas columnas para dejarlas iguales a otros dataframes

df_termino_rol.rename(columns = {'CÓD. CORTE':'COD. CORTE',
                                'CÓD. TRIBUNAL':'COD. TRIBUNAL',
                                'DURACIÓN CAUSA ':'DURACION CAUSA',
                                'MOTIVO DE TÉRMINO':'MOTIVO TERMINO',
                                'FECHA TÉRMINO':'FECHA TERMINO',
                                'MES TÉRMINO':'MES TERMINO',
                                'AÑO TÉRMINO':'AÑO TERMINO',
                                'TOTAL TÉRMINOS':'TOTAL TERMINOS'
                                },
                     inplace = True) 

# Transformamos variables float64 a int16

df_termino_rol['COD. CORTE'] = df_termino_rol['COD. CORTE'].fillna(0).astype(np.int16)
df_termino_rol['COD. TRIBUNAL'] = df_termino_rol['COD. TRIBUNAL'].fillna(0).astype(np.int16)
df_termino_rol['DURACION CAUSA'] = df_termino_rol['DURACION CAUSA'].fillna(0).astype(np.int16)
df_termino_rol['AÑO TERMINO'] = df_termino_rol['AÑO TERMINO'].fillna(0).astype(np.int16)
df_termino_rol['TOTAL TERMINOS'] = df_termino_rol['TOTAL TERMINOS'].fillna(0).astype(np.int8)

# Elimino tildes de las columnas object

cols = df_termino_rol.select_dtypes(include = ["object"]).columns
df_termino_rol[cols] = df_termino_rol[cols].progress_apply(cleandata.elimina_tilde)

# Convertir objetos a fechas
#cols_date = ['FECHA INGRESO', 'FECHA TERMINO']
#df_termino_rol[cols_date] = df_termino_rol[cols_date].progress_apply(convierte_fecha)

df_termino_rol['FECHA INGRESO'] = df_termino_rol['FECHA INGRESO'].progress_apply(cleandata.convierte_fecha)
df_termino_rol['FECHA TERMINO'] = df_termino_rol['FECHA TERMINO'].progress_apply(cleandata.convierte_fecha)

# Elimino espacios en las columnas tipo objetos

df_termino_rol = df_termino_rol.progress_apply(cleandata.elimina_espacios, axis=0)

df_termino_rol['RIT'] = df_termino_rol['RIT'].progress_apply(cleandata.limpia_rit)

# Transformamos en variables categoricas

df_termino_rol['CORTE'] = df_termino_rol['CORTE'].astype('category')
df_termino_rol['MOTIVO TERMINO'] = df_termino_rol['MOTIVO TERMINO'].astype('category')

Dejo solo causas Ordinarias

df_termino_rol['TIPO CAUSA'].unique()

tipo_causa = df_termino_rol[df_termino_rol['TIPO CAUSA']!='Ordinaria']

df_termino_rol.drop(tipo_causa.index, axis=0, inplace=True)

df_termino_rol

# Reset el index para realizar feather
df_termino_rol.reset_index(inplace = True)

# Guardamos dataset como archivo feather
path_interim = "../data/interim/pjud"
os.makedirs(path_interim, exist_ok = True) 

df_termino_rol.to_feather(f'{path_interim}/clean_TerminosRol.feather')
