==> notebooks/3.5-jalvaradoruiz-carga-limpieza-data-inventario.ipynb <==
# Carga Data INVENTARIO

Creación del DataFrame asociado a los Inventarios

Rev: 29-10-2020

import os
import pandas as pd
import numpy as np
from pyarrow import feather
from tqdm import tqdm

from pjud.data import cleandata

path_raw = "../data/raw/pjud" 
archivos = os.listdir(path_raw) 
tqdm.pandas()

# Creacion dataframe con los datos de Inventario Causas en Tramitación Penal de los años 2015 a 2019

dataframes = []

for archivo in archivos:
    if archivo.find("Inventario Causas en Tramitación Penal") != -1:
        df = pd.read_csv(f"{path_raw}/{archivo}", sep = ";", encoding = 'cp850', low_memory = True)
        dataframes.append(df) 
        
df_inventario = pd.concat(dataframes, axis = 0)

df_inventario.columns

df_inventario['SISTEMA'].unique()

Elimino registros de METGE

df_metge = df_inventario[df_inventario['SISTEMA']=='METGE']

df_inventario.drop(df_metge.index, axis=0, inplace=True)

# ESTANDARIZACION DE NOMBRES DE VARIABLES 

df_inventario.rename(columns = {'CÓDIGO CORTE':'COD. CORTE',
                                'CÓDIGO TRIBUNAL':'COD. TRIBUNAL',
                                'CÓDIGO MATERIA':'COD. MATERIA',
                                ' MATERIA':'MATERIA'
                               }, inplace = True)

df_inventario.drop(['SISTEMA'], axis = 'columns', inplace = True)

# TRANSFORMAMOS DE FLOAT A INTEGER

df_inventario['COD. CORTE'] = df_inventario['COD. CORTE'].fillna(0).astype(np.int16)
df_inventario['COD. TRIBUNAL'] = df_inventario['COD. TRIBUNAL'].fillna(0).astype(np.int16)
df_inventario['COD. MATERIA'] = df_inventario['COD. MATERIA'].fillna(0).astype(np.int16)
df_inventario['TOTAL INVENTARIO'] = df_inventario['TOTAL INVENTARIO'].fillna(0).astype(np.int8)

# Transformamos fechas

df_inventario['FECHA INGRESO'] = df_inventario['FECHA INGRESO'].progress_apply(cleandata.convierte_fecha)
df_inventario['FECHA ULT. DILIGENCIA'] = df_inventario['FECHA ULT. DILIGENCIA'].progress_apply(cleandata.convierte_fecha)

# Elimino espacios en las columnas tipo objetos

df_inventario = df_inventario.progress_apply(cleandata.elimina_espacios, axis=0)

# Elimino tildes de las columnas object

cols = df_inventario.select_dtypes(include = ["object"]).columns
df_inventario[cols] = df_inventario[cols].progress_apply(cleandata.elimina_tilde)

# CATEGORIZACION DE VARIABLES

df_inventario['CORTE'] = df_inventario['CORTE'].astype('category')
df_inventario['COMPETENCIA'] = df_inventario['COMPETENCIA'].astype('category')
df_inventario['TIPO ULT. DILIGENCIA'] = df_inventario['TIPO ULT. DILIGENCIA'].astype('category')

Dejo solo causas Ordinarias

df_inventario['TIPO CAUSA'].unique()

tipo_causa = df_inventario[df_inventario['TIPO CAUSA']!='Ordinaria']

df_inventario.drop(tipo_causa.index, axis=0, inplace=True)

# Reset el index para realizar feather
df_inventario.reset_index(inplace = True)

df_inventario

# Guardamos dataset como archivo feather
path_interim = "../data/interim/pjud"
os.makedirs(path_interim, exist_ok = True) 

df_inventario.to_feather(f'{path_interim}/clean_Inventario.feather')
