==> notebooks/3.7-jalvaradoruiz-carga-limpieza-data-duraciones.ipynb <==
# Carga Data DURACIONES

Creación del DataFrame asociado a las duraciones de las causas

Rev: 29-10-2020

import os
import pandas as pd
import numpy as np
from pyarrow import feather
from tqdm import tqdm

from pjud.data import cleandata

path_raw = "../data/raw/pjud" 
archivos = os.listdir(path_raw) 
tqdm.pandas()

# Creacion dataframe con los datos de Duraciones por Rol Penal de los años 2015 a 2019

dataframes = []

for archivo in archivos:
    if archivo.find("Duraciones por Rol Penal") != -1:
        df = pd.read_csv(f"{path_raw}/{archivo}", sep = ";", encoding = 'cp850', dtype = 'unicode', low_memory = True)
        dataframes.append(df) 
        
df_duraciones = pd.concat(dataframes, axis = 0)

df_duraciones.columns

df_duraciones['SISTEMA'].unique()

Elimino causas que no sean SIAGJ

df_no_siagj = df_duraciones[df_duraciones['SISTEMA']!='SIAGJ']

df_duraciones.drop(df_no_siagj.index, axis=0, inplace=True)

df_duraciones.rename(columns = {'CÓD. CORTE':'COD. CORTE',
                                'CÓD. TRIBUNAL':'COD. TRIBUNAL',
                                'DURACIÓN CAUSA ':'DURACIÓN CAUSA',
                                'FECHA TÉRMINO':'FECHA TERMINO',
                                'MOTIVO DE TÉRMINO':'MOTIVO TERMINO',
                                'MES TÉRMINO':'MES TERMINO',
                                'AÑO TÉRMINO':'AÑO TERMINO',
                                'TOTAL TÉRMINOS':'TOTAL TERMINOS'
                               }, inplace = True)

df_duraciones.drop(['N°','SISTEMA'], axis = 'columns', inplace = True)
df_duraciones = df_duraciones.dropna()

# TRANSFORMAMOS DE FLOAT A INTEGER

df_duraciones['COD. CORTE'] = df_duraciones['COD. CORTE'].fillna(0).astype(np.int16)
df_duraciones['COD. TRIBUNAL'] = df_duraciones['COD. TRIBUNAL'].fillna(0).astype(np.int16)
df_duraciones['AÑO TERMINO'] = df_duraciones['AÑO TERMINO'].fillna(0).astype(np.int16)
df_duraciones['TOTAL TERMINOS'] = df_duraciones['TOTAL TERMINOS'].fillna(0).astype(np.int8)

# Transformamos fechas

df_duraciones['FECHA INGRESO'] = df_duraciones['FECHA INGRESO'].progress_apply(cleandata.convierte_fecha)
df_duraciones['FECHA TERMINO'] = df_duraciones['FECHA TERMINO'].progress_apply(cleandata.convierte_fecha)

# Elimino espacios en las columnas tipo objetos

df_duraciones = df_duraciones.progress_apply(cleandata.elimina_espacios, axis=0)

# Elimino tildes 

cols = df_duraciones.select_dtypes(include = ["object"]).columns
df_duraciones[cols] = df_duraciones[cols].progress_apply(cleandata.elimina_tilde) 

# Debo transformar el formato del RIT--AÑO a RIT-AÑO

df_duraciones['RIT'] = df_duraciones['RIT'].progress_apply(cleandata.limpia_rit)

# Categorizacion

df_duraciones['CORTE'] = df_duraciones['CORTE'].astype('category')
df_duraciones['MOTIVO TERMINO'] = df_duraciones['MOTIVO TERMINO'].astype('category')

Dejo solo causas Ordinarias

df_duraciones['TIPO CAUSA'].unique()

tipo_causa = df_duraciones[df_duraciones['TIPO CAUSA']!='Ordinaria']

df_duraciones.drop(tipo_causa.index, axis=0, inplace=True)

# Reset el index para realizar feather

df_duraciones.reset_index(inplace = True)

df_duraciones

# Guardamos dataset como archivo feather
path_interim = "../data/interim/pjud"
os.makedirs(path_interim, exist_ok = True) 

df_duraciones.to_feather(f'{path_interim}/clean_Duraciones.feather')
