==> notebooks/4.1-jalvaradoruiz-construccion-dataset-juzgados-garantia.ipynb <==
# Creación Data referente a Juzgados de Garantía

Se realiza un analisis mediante expresiones regulares del art. 16 del Código Organico de Tribunales (COT). En este artículo se definen las comunas de Chile donde tendrán asiento este tipo de Juzgados y además su territorio de jurisdiccion.

import re
import os
import pandas as pd
import numpy as np

from tqdm import tqdm
from unicodedata import normalize

from pjud.data import cleandata

tqdm.pandas()

path_raw = "./data/raw/cot"

## Articulo 16 COT ref: juzgados de garantías


with open(f'{path_raw}/Juzgados_Garantia.txt', 'r') as file:
    contenido_jg = ''
    for line in file.readlines():
        contenido_jg += line

Se construye una expresion regular para captar una lista con la información que se desea procesar, y así generar un dataframe.

regex_jg=r"(?:(?P<Region>^[\w \']+)\:\n\n)|(?P<JG>^[\w. \-]+)\,\scon\s(?P<Jueces>[\w.\-]+)[a-z\-\s\,]+(?P<Competencia>\.|\s[\w. \-\,]+)"
matches = re.findall(regex_jg, contenido_jg, re.MULTILINE)

#matches

# GENERA UNA ARREGLO DE LISTA DE 4 ELEMENTOS
# ELEMENTO 0 : REGIÓN
# ELEMENTO 1: JUZGADO DE GARANTÍA DE ...
# ELEMENTO 2: Cantidad de Jueces en Letras
# ELEMENTO 3: Competencia (. es lo mismo de 3 0 7) o varias ciudades (separadas por , o y)

#len(matches)

data_jg = []

for item in range(0,len(matches)):
    if matches[item][0] != '':
        region = matches[item][0].upper()
    else:    
        if matches[item][1] != '':
            ciudad = matches[item][1].upper() 
            if ciudad.find("JUZGADO") != -1: 
                juzgado = ciudad
                
            else:
                juzgado = f"JUZGADO DE GARANTIA {ciudad}"

        if matches[item][2] != '':
            cantidad_jueces = cleandata.transforma_numero(matches[item][2])
    
        if matches[item][3] == '.':
            competencia = ciudad
        
        else:  
            if matches[item][3] != '':
                competencia = matches[item][3].upper()
        
        competencia = competencia.replace(" Y ",",")
        competencia = competencia.replace(" E ",",")
        competencia = competencia.replace(".","")
        
        comunas = competencia.split(",")
        
        for comuna in comunas:
            data_jg.append([region,juzgado,ciudad,cantidad_jueces,comuna.strip(),'GARANTIA'])
   
df_juzgados_garantia = pd.DataFrame(data_jg, columns = ['REGION','TRIBUNAL','ASIENTO','JUECES','COMUNA','TIPO JUZGADO'])


# Elimino tildes de las columnas object

cols = df_juzgados_garantia.select_dtypes(include = ["object"]).columns
df_juzgados_garantia[cols] = df_juzgados_garantia[cols].progress_apply(cleandata.elimina_tilde)

df_juzgados_garantia

# Reset el index para realizar feather

df_juzgados_garantia.reset_index(inplace = True)
path_interim = "../data/interim/pjud"
os.makedirs(path_interim, exist_ok = True) 

# Guardamos dataset como archivo feather
df_juzgados_garantia.to_feather(f'{path_interim}/generates_JuzgadosGarantia.feather')
