==> notebooks/4.4-jalvaradoruiz-construccion-dataset-comunas-chile.ipynb <==
# Creación Data referente a Comunas de Chile (fuente : subdere)

Este dataset fue extraido de la Subdere con la division politica de las Comunas, Provincias y Regiones de Chile, acá se consideran las 346 comunas.


import re
import os
import pandas as pd
import numpy as np

from tqdm import tqdm
from unicodedata import normalize

from pjud.data import cleandata

tqdm.pandas()

path_raw = "../data/raw/subdere"

## DataFrame Provincias de Chile  - Fuente: Subdere

df_provincias = pd.read_excel(f"{path_raw}/provinciasChile.xls")

df_provincias

df_provincias.dtypes

# Transformo a mayusculas las columnas de mi interes
df_provincias['Nombre Región'] = df_provincias['Nombre Región'].str.upper()
df_provincias['Nombre Provincia'] = df_provincias['Nombre Provincia'].str.upper()
df_provincias['Nombre Comuna'] = df_provincias['Nombre Comuna'].str.upper()

cols = df_provincias.select_dtypes(include = ["object"]).columns
df_provincias[cols] = df_provincias[cols].progress_apply(cleandata.elimina_tilde)

Acá se debe analiza las comunas que presentar diferentes nombres en los dataset que se estan trabajando, por lo que se procede a unificar la info de comunas con la existente en el Codigo Organico de Tribunales.

# Cambio nombre provincia para coincidir con otro df

df_provincias.loc[df_provincias['Nombre Provincia'] == 'ANTARTICA CHILENA', 'Nombre Provincia'] = 'LA ANTARTICA CHILENA'
df_provincias.loc[df_provincias['Nombre Comuna'] == 'COIHAYQUE', 'Nombre Comuna'] = 'COIHAIQUE'
df_provincias.loc[df_provincias['Nombre Comuna'] == 'PAIGUANO', 'Nombre Comuna'] = 'PAIHUANO'
df_provincias.loc[df_provincias['Nombre Comuna'] == 'TILTIL', 'Nombre Comuna'] = 'TIL TIL'
df_provincias.loc[df_provincias['Nombre Comuna'] == 'EL OLIVAR', 'Nombre Comuna'] = 'OLIVAR'

# Reset el index para realizar feather

df_provincias.reset_index(inplace = True)

path_interim = "../data/interim/subdere"
os.makedirs(path_interim, exist_ok = True) 

# Guardamos dataset como archivo feather
df_provincias.to_feather(f'{path_interim}/generates_Provincias.feather')
