==> notebooks/4.5-jalvaradoruiz-construccion-dataset-censo-2017.ipynb <==
# Data Censo 2017

Se analiza la data del Censo 2017 referente a población por comunas, se establece limpieza y transformación para tener una concordancia con la data del PJUD.

import os
import pandas as pd
import numpy as np
from pyarrow import feather
from datetime import datetime
from datetime import timedelta
from tqdm import tqdm

from pjud.data import cleandata

tqdm.pandas()

path_subdere = "../data/interim/subdere"
path_censo = "../data/raw/censo"

# DATA de COMUNAS OBTENIDAS DEL COT

# Analizo contra los datos extraidos en COT

df_comunas = pd.read_feather(f"{path_subdere}/generates_Provincias.feather")
comunas = df_comunas['Nombre Comuna'].unique()
len(comunas)

## DATA CENSO 2017 : FUENTE INE

df_censo = pd.read_excel(f"{path_censo}/1_1_POBLACION.xls", sheet_name = "Comuna")
df_censo.head()

df_censo.drop(['Unnamed: 0'], axis='columns', inplace=True)
df_censo.drop(0, axis='rows', inplace=True)

old_columns = []
for col in range(1,18):
    old_columns.append(f"Unnamed: {col}")

new_columns = []
for col in old_columns:
    new_columns.append(df_censo[col][1])

columnas = dict(zip(old_columns, new_columns))

df_censo.rename(columns=columnas, inplace=True)

cols = df_censo.select_dtypes(include = ["object"]).columns
df_censo[cols] = df_censo[cols].progress_apply(cleandata.elimina_tilde)

df_censo

seleccion_censo_comunas = df_censo.loc[df_censo["EDAD"].str.contains("Total")]

seleccion_censo_comunas

seleccion_censo_comunas.drop(2, axis='rows', inplace=True)

seleccion_censo_comunas

## Verifico nombres de comunas de ambos dataset

Acá se debe analiza las comunas que presentar diferentes nombres en los dataset que se estan trabajando, por lo que se procede a unificar la info de comunas con la existente en el Codigo Organico de Tribunales.

comunas = df_comunas['Nombre Comuna'].unique()    

for comuna_censo in seleccion_censo_comunas['NOMBRE COMUNA'].unique():
    if not comuna_censo in comunas:
        print(comuna_censo)

 Nota: Tratar de ver otra manera de asignar estos valores 

seleccion_censo_comunas.loc[seleccion_censo_comunas['NOMBRE COMUNA'] == 'COYHAIQUE', 'NOMBRE COMUNA'] = 'COIHAIQUE'
seleccion_censo_comunas.loc[seleccion_censo_comunas['NOMBRE COMUNA'] == 'PAIGUANO', 'NOMBRE COMUNA'] = 'PAIHUANO'
seleccion_censo_comunas.loc[seleccion_censo_comunas['NOMBRE COMUNA'] == 'TILTIL', 'NOMBRE COMUNA'] = 'TIL TIL'
seleccion_censo_comunas.loc[seleccion_censo_comunas['NOMBRE COMUNA'] == 'AYSEN', 'NOMBRE COMUNA'] = 'AISEN'

# Reset el index para realizar feather

seleccion_censo_comunas.reset_index(inplace = True)

path_processed = "../data/processed/censo"
os.makedirs(path_processed, exist_ok = True) 

# Guardamos dataset como archivo feather
seleccion_censo_comunas.to_feather(f'{path_processed}/generates_Censo2017.feather')
