==> notebooks/5.1-jalvaradoruiz-procesamiento-dataset-juzgados-penales.ipynb <==
# Verificación Data de Juzgados

Este notebook tiene como objetivo analizar los dataset creados y corroborar los nombres de Tribunales y que abarquen a las 346 comunas del país.

import re
import os
import pandas as pd
import numpy as np
import warnings

from tqdm import tqdm
from unicodedata import normalize

from pjud.data import cleandata
from pjud.data import transformdata

warnings.filterwarnings(action="ignore")
pd.set_option('display.max_columns', 100)

tqdm.pandas()

path_interim_pjud = "../data/interim/pjud"
path_interim_subdere = "../data/interim/subdere"

### Carga Data Feather referente a tribunales, juzgados y provincias

df_juzgados_garantia = pd.read_feather(f"{path_interim_pjud}/clean_JuzgadosGarantia.feather")
df_tribunal_oral = pd.read_feather(f"{path_interim_pjud}/clean_TribunalOral.feather")
df_juzgados_letras = pd.read_feather(f"{path_interim_pjud}/clean_JuzgadosLetras.feather")
df_provincias = pd.read_feather(f"{path_interim_subdere}/clean_Provincias.feather")

# Existen dos provincias en el listado.
filtro_provincia_top = df_tribunal_oral[df_tribunal_oral['COMUNA'].str.contains("PROVINCIA", case = False)]
filtro_provincia_top # agrego a este dataframe las provincias a explorar por comunas

df_tribunal_oral.drop(filtro_provincia_top.index, axis = 0, inplace = True)

Analizo Dataframe , ya que acá estan los Juzgados de Letras con Competencia Común, y en algunos casos estos pueden No corresponder a Juzgados Penales. Para saber eso, debo verificar que no existan JG en esas comunas, en el caso de existir debo eliminar registro.


df_juzgados_letras

Elimino datos inconcruentes generados y separo las provincias


filtro_provincia_jl = df_juzgados_letras[df_juzgados_letras['COMUNA'].str.contains("PROVINCIA", case = False)]
filtro_provincia_jl

##### Elimino los que estan mal generados o con datos de Provincias


df_juzgados_letras.drop(filtro_provincia_jl.index, axis = 0, inplace = True)

df_juzgados_letras

filtro_provincia_jl

#### EN INDEX 230 , DEJARE SOLO PROVINCIA DE MELIPILLA, DESPUES EXCEPSIONAR CURACAVÍ

filtro_provincia_jl.loc[229,"COMUNA"] = 'PROVINCIA DE MELIPILLA'
filtro_provincia_jl.drop(176, inplace = True)

filtro_provincia_jl

Exploro las provincias y extraigo las comunas -> caso Juzgados Letras 


add_comunas_jl = []

for indice in filtro_provincia_jl.index:
    region = filtro_provincia_jl["REGION"][indice]
    tribunal = filtro_provincia_jl["TRIBUNAL"][indice]
    asiento = filtro_provincia_jl["ASIENTO"][indice]
    jueces = filtro_provincia_jl["JUECES"][indice]
    tipo_juzgado = filtro_provincia_jl["TIPO JUZGADO"][indice]
    
    provincia = filtro_provincia_jl["COMUNA"][indice].split("PROVINCIA DE ")
    comunas_de_provincias = df_provincias.loc[df_provincias["Nombre Provincia"] == provincia[1],"Nombre Comuna"]
    
    for var_comuna in range(len(comunas_de_provincias)):
        comuna = comunas_de_provincias.values[var_comuna]
        add_comunas_jl.append([region,tribunal,asiento,jueces,comuna,tipo_juzgado])


df_nuevas_comunas_jl = pd.DataFrame(add_comunas_jl,columns = ['REGION','TRIBUNAL','ASIENTO','JUECES','COMUNA','TIPO JUZGADO'])     
df_juzgados_letras = df_juzgados_letras.append(df_nuevas_comunas_jl, ignore_index = True)

Exploro las provincias y extraigo las comunas -> Caso Tribunales orales

add_comunas_top = []

for indice in filtro_provincia_top.index:
    region = filtro_provincia_top["REGION"][indice]
    tribunal = filtro_provincia_top["TRIBUNAL"][indice]
    asiento = filtro_provincia_top["ASIENTO"][indice]
    jueces = filtro_provincia_top["JUECES"][indice]
    tipo_juzgado = filtro_provincia_top["TIPO JUZGADO"][indice]
    
    provincia = filtro_provincia_top["COMUNA"][indice].split("PROVINCIA DE ")
    comunas_de_provincias = df_provincias.loc[df_provincias["Nombre Provincia"] == provincia[1],"Nombre Comuna"]
    
    for var_comuna in range(len(comunas_de_provincias)):
        comuna = comunas_de_provincias.values[var_comuna]
        add_comunas_top.append([region,tribunal,asiento,jueces,comuna,tipo_juzgado])


df_nuevas_comunas_top = pd.DataFrame(add_comunas_top,columns = ['REGION','TRIBUNAL','ASIENTO','JUECES','COMUNA','TIPO JUZGADO'])     

df_tribunal_oral = df_tribunal_oral.append(df_nuevas_comunas_top, ignore_index = True)


DE ESTE LISTADO DEBO VERIFICAR LOS QUE TIENEN JUZGADOS DE GARANTIAS Y ELIMINAR DE JUZGADOS DE LETRAS, YA QUE SON LOS JUZGADOS DE GARANTÍA QUE OCUPAN ESA FUNCION EN ESAS COMUNAS

df_juzgados_garantia

Acá debo eliminar los juzgados de letras que estan cubiertos por la competencia de un juzgado de Garantía

juzgados_garantias = df_juzgados_garantia['COMUNA'].unique().tolist()
drop_jl = []

for jl in df_juzgados_letras.index:
    if df_juzgados_letras['COMUNA'][jl] in juzgados_garantias:
        drop_jl.append(jl)

df_juzgados_letras.drop(drop_jl, axis = 0, inplace = True)

df_juzgados_letras

df_juzgados_penales = pd.concat([df_juzgados_garantia,df_juzgados_letras], join = "inner")

df_juzgados_penales.reset_index(inplace = True)

df_juzgados_penales.count()

df_tribunal_oral.count()

Verifico cuales comunas no existen en juzgados penales --> Casos especiales

for indice in df_provincias.index:
    if not df_provincias['Nombre Comuna'][indice] in df_juzgados_penales['COMUNA'].unique():
        print(df_provincias['Nombre Comuna'][indice])

Verifico cuales comunas no existen en Tribunales orales --> Casos especiales

for indice in df_provincias.index:
    if not df_provincias['Nombre Comuna'][indice] in df_tribunal_oral['COMUNA'].unique():
        print(df_provincias['Nombre Comuna'][indice])

El caso de Alto Hospicio es particular, no aparece en el COT, ya que fue parte de una actualizacion de la Ley y fue sumado como Juzgado especial. Debemos agregarlo a Juzgados Penales

alto_hospicio = [('REGION DE TARAPACA','JUZGADO DE LETRAS Y GARANTIA ALTO HOSPICIO','ALTO HOSPICIO','4',
                 'ALTO HOSPICIO','LETRAS Y GARANTIA')]

df_alto_hospicio = pd.DataFrame(alto_hospicio, columns = ['REGION','TRIBUNAL','ASIENTO','JUECES','COMUNA','TIPO JUZGADO']) 
df_juzgados_penales = pd.concat([df_juzgados_penales, df_alto_hospicio], join = "inner")
df_juzgados_penales['JUECES'] = df_juzgados_penales['JUECES'].fillna(0).astype(np.int8)
df_juzgados_penales.reset_index(drop = True)

# Transformo las variables para que queden igual en los 3 datasets

df_juzgados_penales['REGION'] = df_juzgados_penales['REGION'].progress_apply(cleandata.separa_regiones)
df_tribunal_oral['REGION'] = df_tribunal_oral['REGION'].progress_apply(cleandata.separa_regiones)

df_juzgados_penales['ASIENTO'] = df_juzgados_penales['ASIENTO'].progress_apply(cleandata.transforma_asiento)
df_tribunal_oral['ASIENTO'] = df_tribunal_oral['ASIENTO'].progress_apply(cleandata.transforma_asiento)

# Transformo valores de Regiones a similar a BBDD Pjud

regiones_subdere = df_provincias["Nombre Región"].unique()
regiones_top = df_tribunal_oral["REGION"].unique()

for region in regiones_subdere:
    for region_top in regiones_top:
        if region_top.find(region) != -1:
            df_provincias['Nombre Región'] = df_provincias['Nombre Región'].replace(region, region_top) 
    
df_provincias['Nombre Región'] = df_provincias['Nombre Región'].progress_apply(cleandata.separa_regiones)

# Creare dataframe con tribunales ORales- Garantia y Letras 

top =  df_tribunal_oral[['REGION','TRIBUNAL','ASIENTO','COMUNA','JUECES','TIPO JUZGADO']] 
jg = df_juzgados_penales[['REGION','TRIBUNAL','ASIENTO','COMUNA','JUECES','TIPO JUZGADO']] 

df_listado_tribunales = pd.concat([top, jg], join = "inner")

df_tribunal_oral['TRIBUNAL'] = df_tribunal_oral['TRIBUNAL'].progress_apply(cleandata.cambio_nombre_juzgados)
df_juzgados_penales['TRIBUNAL'] = df_juzgados_penales['TRIBUNAL'].progress_apply(cleandata.cambio_nombre_juzgados)
df_listado_tribunales['TRIBUNAL'] = df_listado_tribunales['TRIBUNAL'].progress_apply(cleandata.cambio_nombre_juzgados)

## GUARDAMOS DATA EN FEATHER

# Reset el index para realizar feather

df_tribunal_oral.reset_index(inplace = True)
df_juzgados_penales.reset_index(inplace = True)
df_listado_tribunales.reset_index(inplace = True)
df_provincias.reset_index(inplace = True)

# Directorio donde se guardaran archivos feather

path_processed_subdere = "../data/processed/subdere"
os.makedirs(path_processed_subdere, exist_ok = True) 
df_provincias.to_feather(f'{path_processed_subdere}/processes_DataRegiones.feather')

path_processed = "../data/processed/pjud"
os.makedirs(path_processed, exist_ok = True) 
df_listado_tribunales.to_feather(f'{path_processed}/generates_ListadoTribunales.feather')
df_tribunal_oral.to_feather(f'{path_processed}/processes_TribunalesOrales.feather')
df_juzgados_penales.to_feather(f'{path_processed}/processes_JuzgadosPenales.feather')
