==> notebooks/5.2-jalvaradoruiz-procesamiento-dataset-materia.ipynb <==
# Generación de Data Única referente a Materia

Unificacion de la data de ingresos y terminos por Materia penal, además de transformación de algunos datos para normalizar nombres en columnas

import pandas as pd
import numpy as np
import os

from tqdm import tqdm
from unicodedata import normalize

from pjud.data import cleandata
from pjud.data import transformdata

tqdm.pandas()

path_interim = "../data/interim/pjud"
path_processed = "../data/processed/pjud"

# DATA de TRIBUNALES OBTENIDAS DEL COT

# Analizo contra los datos extraidos en COT
df_tribunales = pd.read_feather(f"{path_processed}/generates_ListadoTribunales.feather")
tribunales = df_tribunales['TRIBUNAL'].unique()

# DATAS INGRESO Y TERMINO MATERIA

### DATA INGRESO X MATERA

df_ingresos_materia = pd.read_feather(f"{path_interim}/clean_IngresosMateria.feather")
df_ingresos_materia.head()


# Normalizo nombres

df_ingresos_materia['TRIBUNAL'] = df_ingresos_materia['TRIBUNAL'].progress_apply(cleandata.cambio_nombre_juzgados)

# Verifico si coinciden los nombres e imprimo los que estan en conflicto

for tribunal in df_ingresos_materia['TRIBUNAL'].unique():
    if not tribunal in tribunales:
        print(tribunal)

### TERMINO X MATERIA

df_termino_materia = pd.read_feather(f"{path_interim}/clean_TerminosMateria.feather")
df_termino_materia.head()

# Normalizo nombres

df_termino_materia['TRIBUNAL'] = df_termino_materia['TRIBUNAL'].progress_apply(cleandata.cambio_nombre_juzgados)

# Verifico si coinciden los nombres e imprimo los que estan en conflicto

for tribunal in df_termino_materia['TRIBUNAL'].unique():
    if not tribunal in tribunales:
        print(tribunal)

## Data relacionada a Delitos

path = "../data/processed/delitos"
df_delitos = pd.read_feather(f"{path}/clean_Delitos.feather")
df_delitos.head()

codigos_cot = df_delitos['COD. MATERIA'].unique()
print(f"Existen {len(codigos_cot)} codigos de delitos")

## Delitos en INGRESOS MATERIA

codigos_delitos = df_ingresos_materia['COD. MATERIA'].unique()

len(codigos_delitos)

codigos_no_encontrados = []
for delito in codigos_delitos:
    if not delito in codigos_cot:
        codigos_no_encontrados.append(delito)

codigos_no_encontrados

codigos_sin_registro = df_ingresos_materia[df_ingresos_materia['COD. MATERIA'].isin(codigos_no_encontrados)]

codigos_sin_registro['MATERIA'].unique()

# Analizar caso a caso:

#### Caso Codigo 13001 -> LESIONES LEVES
Este codigo debe tratarse como 13036 identificado como: LESIONES LEVES 494 Nº 5 CÓDIGO PENAL

df_ingresos_materia['COD. MATERIA'] = df_ingresos_materia['COD. MATERIA'].replace(13001,13036)

#### Caso Codigo -31075

filtro_codigo_31075 = df_ingresos_materia[df_ingresos_materia['COD. MATERIA']==-31075]
filtro_codigo_31075

No corresponde a delito sino a sanción interna de gendarmeria, por lo que se eliminarán


df_ingresos_materia.drop(filtro_codigo_31075.index, axis = 0, inplace = True)

### Caso Codigo 0

filtro_codigo_0 = df_ingresos_materia[df_ingresos_materia['COD. MATERIA']==0]
filtro_codigo_0

#### valores Null

df_ingresos_materia.drop(filtro_codigo_0.index, axis = 0, inplace = True)

### Codigos 4011, 4014, 4013, 4010
codigos No vigentes, mal cargados en sistemas 

filtro_codigo_4011 = df_ingresos_materia[df_ingresos_materia['COD. MATERIA']==4011]
filtro_codigo_4014 = df_ingresos_materia[df_ingresos_materia['COD. MATERIA']==4014]
filtro_codigo_4013 = df_ingresos_materia[df_ingresos_materia['COD. MATERIA']==4013]
filtro_codigo_4010 = df_ingresos_materia[df_ingresos_materia['COD. MATERIA']==4010]

df_ingresos_materia.drop(filtro_codigo_4011.index, axis = 0, inplace = True)
df_ingresos_materia.drop(filtro_codigo_4014.index, axis = 0, inplace = True)
df_ingresos_materia.drop(filtro_codigo_4013.index, axis = 0, inplace = True)
df_ingresos_materia.drop(filtro_codigo_4010.index, axis = 0, inplace = True)

pd.to_datetime(df_ingresos_materia['FECHA INGRESO'])

## DELITOS EN TERMINO MATERIA

codigos_delitos = df_termino_materia['COD. MATERIA'].unique()

len(codigos_delitos)

codigos_no_encontrados = []
for delito in codigos_delitos:
    if not delito in codigos_cot:
        codigos_no_encontrados.append(delito)

codigos_no_encontrados

codigos_sin_registro = df_termino_materia[df_termino_materia['COD. MATERIA'].isin(codigos_no_encontrados)]

codigos_sin_registro['MATERIA'].unique()

# Analizar caso a caso:

## Caso 13001 -- > LESIONES LEVES

df_termino_materia['COD. MATERIA'] = df_termino_materia['COD. MATERIA'].replace(13001,13036)

### Casos 4014, 4013, 4011, -31075, 4010, 0, 4007

filtro_codigo_4011 = df_termino_materia[df_termino_materia['COD. MATERIA']==4011]
filtro_codigo_4014 = df_termino_materia[df_termino_materia['COD. MATERIA']==4014]
filtro_codigo_4013 = df_termino_materia[df_termino_materia['COD. MATERIA']==4013]
filtro_codigo_4010 = df_termino_materia[df_termino_materia['COD. MATERIA']==4010]
filtro_codigo_31075 = df_termino_materia[df_termino_materia['COD. MATERIA']==31075]
filtro_codigo_4007 = df_termino_materia[df_termino_materia['COD. MATERIA']==4007]
filtro_codigo_0 = df_termino_materia[df_termino_materia['COD. MATERIA']==0]

df_termino_materia.drop(filtro_codigo_4011.index, axis = 0, inplace = True)
df_termino_materia.drop(filtro_codigo_4014.index, axis = 0, inplace = True)
df_termino_materia.drop(filtro_codigo_4013.index, axis = 0, inplace = True)
df_termino_materia.drop(filtro_codigo_4010.index, axis = 0, inplace = True)
df_termino_materia.drop(filtro_codigo_31075.index, axis = 0, inplace = True)
df_termino_materia.drop(filtro_codigo_4007.index, axis = 0, inplace = True)
df_termino_materia.drop(filtro_codigo_0.index, axis = 0, inplace = True)

pd.to_datetime(df_termino_materia['FECHA INGRESO'])
pd.to_datetime(df_termino_materia['FECHA TERMINO'])

# Creacion Consolidado por Materia Penal

Comenzaremos el analisis de la data consolidando la info en un unico DataSet, para ello vamos a considerar los siguientes puntos:


- DATA:  Se considerará toda las causas que ingresaron al Sistema desde 2015 al 2019, esto ya que se cuenta además con la data de las audiencias en ese período.


- No se consideran causas anteriores a este período.

- Se considerarán Causas sin termino.

df_ingresos_materia.sort_values('FECHA INGRESO')

df_termino_materia.sort_values('FECHA INGRESO')

##### Verifico si existen Causas sin Fechas de inicio

filtro = df_ingresos_materia[df_ingresos_materia['FECHA INGRESO'].isnull()]

print(f"Existen: {len(filtro)} causas con Fechas Ingreso Null")

##### Verifico si existen Causas con Fechas de inicio con NULL en Termino_Materia

filtro_null = df_termino_materia[df_termino_materia['FECHA INGRESO'].isnull()]
print(f"Existen {len(filtro_null)} causas con Fecha Ingreso Null")

Elimino las filas con valores nulos

df_termino_materia.drop(filtro_null.index, axis = 0, inplace = True)

Verifico las causas con Fecha Ingreso anterior a 01-01-2015

filtro_fechas = df_termino_materia[df_termino_materia['FECHA INGRESO'] <= '2014-12-31']
print(f"Existen {len(filtro_fechas)} causas con Fecha Ingreso anteriores al 31-12-2014 que no deben ser considerada")

Elimino las causas con Fecha de Ingreso anterior a 31-12-2014

df_termino_materia.drop(filtro_fechas.index, axis = 0, inplace = True)

df_termino_materia

Verifico si existen Causas con duracion negativo, esto se debe a error en Sistema y debe cambiar el orden.

df_termino_materia[df_termino_materia['DURACION CAUSA']<0]

df_termino_materia = df_termino_materia.progress_apply(transformdata.fechas_cambiadas, axis=1)

df_termino_materia[df_termino_materia['DURACION CAUSA']<0]

df_ingresos_materia['AÑO INGRESO'].value_counts()

df_termino_materia[df_termino_materia['FECHA INGRESO']<'01-01-2015']

## CARGA DATA A ARCHIVOS FUTHER

# Directorio donde se guardaran archivos feather
df_ingresos_materia.reset_index(inplace=True, drop=True)
df_termino_materia.reset_index(inplace=True, drop=True)
os.makedirs(path_processed, exist_ok = True)  

# Guardamos dataset como archivo feather

df_ingresos_materia.to_feather(f'{path_processed}/processes_IngresosMateria.feather')
df_termino_materia.to_feather(f'{path_processed}/processes_TerminosMateria.feather')
