==> notebooks/5.3-jalvaradoruiz-procesamiento-dataset-rol.ipynb <==
# Generación de Data Única referente a Rol

Unificacion de la data de ingresos y terminos por Rol, además de transformación de algunos datos para normalizar nombres en columnas

import pandas as pd
import numpy as np
import os
from tqdm import tqdm
from unicodedata import normalize

from pjud.data import cleandata
from pjud.data import transformdata

tqdm.pandas()

path_interim = "../data/interim/pjud"
path_processed = "../data/processed/pjud"

# DATA de TRIBUNALES OBTENIDAS DEL COT

# Analizo contra los datos extraidos en COT
df_tribunales = pd.read_feather(f"{path_processed}/generates_ListadoTribunales.feather")
tribunales = df_tribunales['TRIBUNAL'].unique()

tribunales

# DATA INGRESO Y TERMINO ROL PENAL

### INGRESOS X ROL

df_ingresos_rol = pd.read_feather(f"{path_interim}/clean_IngresosRol.feather")
df_ingresos_rol.sort_values('FECHA INGRESO')

df_ingresos_rol['TRIBUNAL'].unique()

# Normalizo nombres

df_ingresos_rol['TRIBUNAL'] = df_ingresos_rol['TRIBUNAL'].progress_apply(cleandata.cambio_nombre_juzgados)

# Verifico si coinciden los nombres e imprimo los que estan en conflicto

for tribunal in df_ingresos_rol['TRIBUNAL'].unique():
    if not tribunal in tribunales:
        print(tribunal)

df_ingresos_rol['TRIBUNAL'].unique()

filtro_null = df_ingresos_rol[df_ingresos_rol['FECHA INGRESO'].isnull()]
print(f"Existen {len(filtro_null)} causas con Fechas Null")

### TERMINO X ROL

df_termino_rol = pd.read_feather(f"{path_interim}/clean_TerminosRol.feather")
df_termino_rol.sort_values('FECHA INGRESO')

filtro_null = df_termino_rol[df_termino_rol['FECHA INGRESO'].isnull()]
print(f"Existen {len(filtro_null)} causas con ingresos NULL")

df_termino_rol.drop(filtro_null.index, axis=0, inplace=True)

filtro_fecha = df_termino_rol[df_termino_rol['FECHA INGRESO'] <='2014-12-31']
print(f"Existen {len(filtro_fecha)} causas con ingresos anteriores a 01-01-2015")

df_termino_rol.drop(filtro_fecha.index, axis=0, inplace=True)

# Normalizo nombres

df_termino_rol['TRIBUNAL'] = df_termino_rol['TRIBUNAL'].progress_apply(cleandata.cambio_nombre_juzgados)

# Verifico si coinciden los nombres e imprimo los que estan en conflicto

for tribunal in df_termino_rol['TRIBUNAL'].unique():
    if not tribunal in tribunales:
        print(tribunal)

df_termino_rol[df_termino_rol['DURACION CAUSA']<0]

df_termino_rol = df_termino_rol.progress_apply(transformdata.fechas_cambiadas, axis=1)

df_termino_rol[df_termino_rol['DURACION CAUSA']<0]

df_ingresos_rol['AÑO INGRESO'].value_counts()

len(df_ingresos_rol)

## CARGA DATA A ARCHIVOS FUTHER

# Directorio donde se guardaran archivos feather
df_ingresos_rol.reset_index(inplace=True, drop=True)
df_termino_rol.reset_index(inplace=True, drop=True)
os.makedirs(path_processed, exist_ok = True)  

# Guardamos dataset como archivo feather

df_ingresos_rol.to_feather(f'{path_processed}/processes_IngresosRol.feather')
df_termino_rol.to_feather(f'{path_processed}/processes_TerminosRol.feather')
