==> notebooks/5.4-jalvaradoruiz-procesamiento-dataset-audiencias.ipynb <==
# Generación de Data Única referente a Audiencias

Transformación de algunos datos para normalizar nombres en columnas

import pandas as pd
import numpy as np
import os
from tqdm import tqdm
from unicodedata import normalize


from pjud.data import cleandata
from pjud.data import transformdata

tqdm.pandas()

path_interim = "../data/interim/pjud"
path_processed = "../data/processed/pjud"

# DATA de TRIBUNALES OBTENIDAS DEL COT

# Analizo contra los datos extraidos en COT
df_tribunales = pd.read_feather(f"{path_processed}/generates_ListadoTribunales.feather")
tribunales = df_tribunales['TRIBUNAL'].unique()

# DATA AUDIENCIAS


df_audiencias = pd.read_feather(f"{path_interim}/clean_Audiencias.feather")
df_audiencias.sort_values('FECHA PROGRAMACION AUDIENCIA')

df_audiencias[df_audiencias['FECHA PROGRAMACION AUDIENCIA'].isnull()]

Existen audiencias soin FECHA PROGRAMACION AUDIENCIA, pero se cuenta con la fecha de la AUDIENCIA y además con DIAS AGENDAMIENTO, por lo que se puede acceder.

df_audiencias = df_audiencias.progress_apply(transformdata.fecha_programada, axis=1)

# RAW: se deben considerar todas las audiencias.

filtro_fecha = df_audiencias[df_audiencias['FECHA PROGRAMACION AUDIENCIA']<='2014-12-31']
print(f"Existen {len(filtro_fecha)} AUDIENCIAS programadas antes del 01-01-2015")

# RAW: se deben considerar todas las audiencias.

df_audiencias.drop(filtro_fecha.index, axis=0, inplace=True)

# Normalizo nombres

df_audiencias['TRIBUNAL'] = df_audiencias['TRIBUNAL'].progress_apply(cleandata.cambio_nombre_juzgados)

# Verifico si coinciden los nombres e imprimo los que estan en conflicto

for tribunal in df_audiencias['TRIBUNAL'].unique():
    if not tribunal in tribunales:
        print(tribunal)

Verifico si existen audiencias de causas del período anterior a 2015

# RAW: se deben considerar todas las audiencias.

años = range(2000,2015)
fuera_rango = []

for año in range(len(años)):
    var = "-" + str(años[año])
    fuera_rango.append(df_audiencias[df_audiencias['RIT'].str.contains(var)])


# RAW: se deben considerar todas las audiencias.

df_eliminar = pd.concat(fuera_rango, axis=0)

# RAW: se deben considerar todas las audiencias.
df_eliminar.sort_values("RIT")

# RAW: se deben considerar todas las audiencias.
print(f"Existen {len(df_eliminar)} audiencias del periodo anterior a 2015, estas se deben eliminar del dataset Audiencias")

# RAW: se deben considerar todas las audiencias.
df_audiencias.drop(df_eliminar.index, axis=0, inplace=True)

#df_audiencias['FECHA PROGRAMACION AUDIENCIA'].sort_values()

## CARGA DATA A ARCHIVOS FUTHER

# Directorio donde se guardaran archivos feather
df_audiencias.reset_index(inplace = True)
os.makedirs(path_processed, exist_ok = True)  

# Guardamos dataset como archivo feather
df_audiencias.to_feather(f'{path_processed}/processes_Audiencias.feather')
