==> notebooks/5.5-jalvaradoruiz-procesamiento-dataset-inventarios.ipynb <==
# Generación de Data Única referente a Inventario

Transformación de algunos datos para normalizar nombres en columnas

import pandas as pd
import numpy as np
import os
from tqdm import tqdm
from unicodedata import normalize

from pjud.data import cleandata

tqdm.pandas()

path_interim = "../data/interim/pjud"
path_processed = "../data/processed/pjud"

# DATA de TRIBUNALES OBTENIDAS DEL COT

# Analizo contra los datos extraidos en COT
df_tribunales = pd.read_feather(f"{path_processed}/generates_ListadoTribunales.feather")
tribunales = df_tribunales['TRIBUNAL'].unique()

# DATA INVENTARIO

df_inventario = pd.read_feather(f"{path_interim}/clean_Inventario.feather")
df_inventario.sort_values('FECHA INGRESO')

# RAW: se deben considerar todas.
filtro_fecha = df_inventario[df_inventario['FECHA INGRESO']<='2014-12-31']
print(f"Existen {len(filtro_fecha)} causas ingresadas antes del 01-01-2015")

df_inventario.drop(filtro_fecha.index, axis=0, inplace=True)

filtro_null = df_inventario[df_inventario['FECHA INGRESO'].isnull()]
print(f"Existen {len(filtro_null)} causas sin fecha de ingresos")

df_inventario.drop(filtro_null.index, axis=0, inplace=True)

# Normalizo nombres

df_inventario['TRIBUNAL'] = df_inventario['TRIBUNAL'].progress_apply(cleandata.cambio_nombre_juzgados)

# Verifico si coinciden los nombres e imprimo los que estan en conflicto

for tribunal in df_inventario['TRIBUNAL'].unique():
    if not tribunal in tribunales:
        print(tribunal)

## Data relacionada a delitos

path = "../data/processed/delitos"
df_delitos = pd.read_feather(f"{path}/clean_Delitos.feather")
df_delitos.head()

codigos_cot = df_delitos['COD. MATERIA'].unique()
print(f"Existen {len(codigos_cot)} codigos de delitos")

## Delitos en INVENTARIO

codigos_delitos = df_inventario['COD. MATERIA'].unique()

len(codigos_delitos)

codigos_no_encontrados = []
for delito in codigos_delitos:
    if not delito in codigos_cot:
        codigos_no_encontrados.append(delito)

codigos_no_encontrados

codigos_sin_registro = df_inventario[df_inventario['COD. MATERIA'].isin(codigos_no_encontrados)]

# Analizar caso a caso:

#### Caso Codigo 13001 -> LESIONES LEVES
Este codigo debe tratarse como 13036 identificado como: LESIONES LEVES 494 Nº 5 CÓDIGO PENAL

df_inventario['COD. MATERIA'] = df_inventario['COD. MATERIA'].replace(13001,13036)

### Codigos 4011, 4014, 4013, 4010, -31075, 0
codigos No vigentes, mal cargados en sistemas 

filtro_codigo_4011 = df_inventario[df_inventario['COD. MATERIA']==4011]
filtro_codigo_4014 = df_inventario[df_inventario['COD. MATERIA']==4014]
filtro_codigo_4013 = df_inventario[df_inventario['COD. MATERIA']==4013]
filtro_codigo_4010 = df_inventario[df_inventario['COD. MATERIA']==4010]
filtro_codigo_31075 = df_inventario[df_inventario['COD. MATERIA']==31075]
filtro_codigo_0 = df_inventario[df_inventario['COD. MATERIA']==0]

df_inventario.drop(filtro_codigo_4011.index, axis = 0, inplace = True)
df_inventario.drop(filtro_codigo_4014.index, axis = 0, inplace = True)
df_inventario.drop(filtro_codigo_4013.index, axis = 0, inplace = True)
df_inventario.drop(filtro_codigo_4010.index, axis = 0, inplace = True)
df_inventario.drop(filtro_codigo_31075.index, axis = 0, inplace = True)
df_inventario.drop(filtro_codigo_0.index, axis = 0, inplace = True)

## CARGA DATA A ARCHIVOS FUTHER

# Directorio donde se guardaran archivos feather
df_inventario.reset_index(inplace=True, drop=True)
os.makedirs(path_processed, exist_ok = True)  
# Guardamos dataset como archivo feather
df_inventario.to_feather(f'{path_processed}/processes_Inventario.feather')
