==> notebooks/5.7-jalvaradoruiz-procesamiento-dataset-cortes.ipynb <==
# Generación de Data Única referente a Cortes , tribunales y Regiones

Transformación de algunos datos para normalizar nombres en columnas, además de unificacion por comunas con data de Censo

import pandas as pd
import numpy as np
import os

from tqdm import tqdm
from unicodedata import normalize

### Analizo Data de Tribunales y Cortes de Apelaciones

path_processed = "../data/processed/pjud"
path_censo = "../data/processed/censo"

df_regiones = pd.read_feather(f"{path_processed}/processes_TerminosRol.feather")
df_tribunales = pd.read_feather(f"{path_processed}/generates_ListadoTribunales.feather")
df_censo = pd.read_feather(f"{path_censo}/generates_Censo2017.feather")


df_tribunales

data_cortes_apelaciones = pd.unique(df_regiones[['CORTE','TRIBUNAL']].values.ravel())

data_cortes_apelaciones

# Extraigo Cortes de Apelaciones asociadas a Juzgados desde este DataSet
data_cortes_apelaciones = pd.unique(df_regiones[['CORTE','TRIBUNAL']].values.ravel())
cortes_tribunales = []

for data in range(len(data_cortes_apelaciones)):  
    if not data_cortes_apelaciones[data].find('C.A.') == -1:
        corte_apelacion = data_cortes_apelaciones[data]
    else:
        tribunal = data_cortes_apelaciones[data]
        if tribunal.find("TRIBUNAL") != -1:
            separa_ciudad = tribunal.split("PENAL ")
        else:
            separa_ciudad = tribunal.split("GARANTIA ")
        ciudad = separa_ciudad[1]
        cortes_tribunales.append([corte_apelacion,ciudad,tribunal])

#cortes_tribunales

lista_cortes = []

for trib in df_tribunales.index:
    for indice in range(len(cortes_tribunales)):
        if df_tribunales['TRIBUNAL'][trib] == cortes_tribunales[indice][2]:
            corte = cortes_tribunales[indice][0]
            lista_cortes.append(corte)
            break

df_tribunales_cortes = df_tribunales.copy()
df_tribunales_cortes['CORTE'] = lista_cortes

df_tribunales

df_tribunales_cortes

df_tribunales_cortes.query("CORTE == 'C.A. DE SAN MIGUEL'")

df_censo.columns

poblacion = []

for trib in df_tribunales.index:
    for indice in df_censo.index:
        if df_tribunales['COMUNA'][trib] == df_censo['NOMBRE COMUNA'][indice]:
            censado = df_censo['TOTAL POBLACIÓN EFECTIVAMENTE CENSADA'][indice]
            hombres = df_censo['HOMBRES '][indice]
            mujeres = df_censo['MUJERES'][indice]
            urbana = df_censo['TOTAL ÁREA URBANA'][indice]
            rural = df_censo['TOTAL ÁREA RURAL'][indice]
            poblacion.append([censado, hombres, mujeres, urbana, rural])
            break

df_poblacion = pd.DataFrame(poblacion, columns = ['POBLACION', 'HOMBRES', 'MUJERES', 'URBANO', 'RURAL'])

df_poblacion

df_tribunales_poblacion = pd.concat([df_tribunales_cortes, df_poblacion.reindex(df_tribunales_cortes.index)], axis=1)

df_tribunales_poblacion

Creo un dataset con informacion de poblacion que abarca cada tribunal !!!

tribunales = df_tribunales['TRIBUNAL'].unique()
poblacion_tribunal = []

for trib in range(len(tribunales)):  
    poblacion = 0
    hombres = 0
    mujeres = 0
    urbana = 0
    rural = 0
    comunas = []
    for indice in df_tribunales_poblacion.index:
        if tribunales[trib] == df_tribunales_poblacion['TRIBUNAL'][indice]:
            region = df_tribunales_poblacion['REGION'][indice]
            corte = df_tribunales_poblacion['CORTE'][indice]
            tribunal = df_tribunales_poblacion['TRIBUNAL'][indice]
            poblacion = int(df_tribunales_poblacion['POBLACION'][indice]) + poblacion
            hombres = int(df_tribunales_poblacion['HOMBRES'][indice]) + hombres
            mujeres = int(df_tribunales_poblacion['MUJERES'][indice]) + mujeres
            urbana = int(df_tribunales_poblacion['URBANO'][indice]) + urbana
            rural = int(df_tribunales_poblacion['RURAL'][indice]) + rural
            comunas.append(df_tribunales_poblacion['COMUNA'][indice])
            
    poblacion_tribunal.append([region, corte, tribunal, poblacion, hombres, mujeres, urbana, rural, comunas])
    
    
columnas = ['REGION', 'CORTE', 'TRIBUNAL', 'POBLACION', 'HOMBRES', 'MUJERES', 'URBANO', 'RURAL', 'COMUNAS']
df_poblacion_jurisdiccion = pd.DataFrame(poblacion_tribunal, columns = columnas)

df_poblacion_jurisdiccion

Agregare data de cantidad de jueces a esta Data y Asientio de cada Tribunal.

dotacion = []
i=0
for indice in df_poblacion_jurisdiccion.index:
    for trib in df_tribunales.index:
        if df_poblacion_jurisdiccion['TRIBUNAL'][indice] == df_tribunales['TRIBUNAL'][trib]:
            jueces = df_tribunales['JUECES'][trib]
            asiento = df_tribunales['ASIENTO'][trib]
            tipo = df_tribunales['TIPO JUZGADO'][trib]
            row = [jueces, asiento, tipo]
            dotacion.append(row)
            break
columnas = ['JUECES','ASIENTO','TIPO JUZGADO']
df_anexo = pd.DataFrame(dotacion, columns=columnas)

df_poblacion_jurisdiccion = pd.concat([df_poblacion_jurisdiccion,df_anexo], axis=1)

df_poblacion_jurisdiccion

df_poblacion_jurisdiccion.query("CORTE == 'C.A. DE SAN MIGUEL'")

## CARGA DATA A ARCHIVOS FUTHER

# Reset el index para realizar feather
df_tribunales_cortes.reset_index(inplace = True)
df_tribunales_poblacion.reset_index(inplace = True)
df_poblacion_jurisdiccion.reset_index(inplace=True)
# Directorio donde se guardaran archivos feather
os.makedirs(path_processed, exist_ok = True) 
df_tribunales_cortes.to_feather(f'{path_processed}/processes_ListadoTribunalesyCortes.feather')
df_tribunales_poblacion.to_feather(f'{path_processed}/processes_DataConsolidada_Poblacion_Tribunales.feather')
df_poblacion_jurisdiccion.to_feather(f'{path_processed}/processes_DataConsolidada_Poblacion_Jurisdiccion.feather')
