==> notebooks/6.1-jalvaradoruiz-consolidacion-dataset-materia.ipynb <==
# Creacion Consolidado por Materia Penal

Comenzaremos el analisis de la data consolidando la info en un unico DataSet, para ello vamos a considerar los siguientes puntos:


- DATA:  Se considerará toda las causas que ingresaron al Sistema desde 2015 al 2019, esto ya que se cuenta además con la data de las audiencias en ese período.


- No se consideran causas anteriores a este período.

- Se considerarán Causas sin termino.

import pandas as pd
import numpy as np
import os
import warnings

from tqdm import tqdm
from unicodedata import normalize
from pjud.data import cleandata
from pjud.data import transformdata

warnings.filterwarnings(action="ignore")
tqdm.pandas()

pd.set_option('display.max_columns', 100)

path_processed = "../data/processed/pjud"

#### Paso 1 : Consolidacion de Dataset x Materia y x rol

df_ingresos_materia = pd.read_feather(f"{path_processed}/processes_IngresosMateria.feather")
print(f"{len(df_ingresos_materia)} causas en el dataset de ingresos por materia")

df_termino_materia = pd.read_feather(f"{path_processed}/processes_TerminosMateria.feather")
print(f"Existen : {len(df_termino_materia)} causas en el dataset de termino por materia")

Analizamos la data, uniendo ambos dataset.

df_fulldata_materia = pd.merge(df_ingresos_materia, df_termino_materia, how='outer', on=['COD. TRIBUNAL','RIT','COD. MATERIA'])
print(f"Existen: {len(df_fulldata_materia)} causas en total uniendo ambos dataset, aca existen datos sin registro de ingreso")

df_fulldata_materia['AÑO INGRESO'].value_counts()

### Analisis de la Data Causas por Materia

De estos dos dataset debo considerar:

- Caso 1: el dataset df_causas_terminadas completo
- Caso 2: Debo agregar además las causas que tengan Fecha Ingreso_x pero que aún no terminan.
- Caso 3: Debo analizar las causas que no tienen Fecha Ingreso_x pero si tienen Fecha Ingreso_y y Fecha Ingreso_y, ya que deben ser las causas que no fueron tomadas por la consulta de Ingresos. Acá se debe reemplazar los datos obtenidos en Termino Materia a Ingresos.

Luego de tener estos tres dataset debo concatenarlos

df_fulldata_materia

df_fulldata_materia.columns

columnas_drop = ['index_x', 'index_y', 'MES INGRESO', 'MES TERMINO']

df_fulldata_materia.drop(columnas_drop, axis = 'columns', inplace = True)

Aca podemos completar algunos datos faltantes utilizando los datos del dataset INGRESOS:

    - FECHA INGRESO_x en Termino
    - MATERIA_x
o

    - FECHA INGRESO_y en Ingreso
    - MATERIA_y
    
Esto debido a que los datos faltantes se deben principalmente a formas de tramitar las causas en los diferentes tribunales. Esto despues de analizar una cantidad importante de Causas del Tribunal Oral en lo penal de Punta Arenas , cod: Tribunal 991 y junto a Maritza Romero, Jefa de Causas y abogada donde se concluyo:

"Bien. Llevo 5 causas revisadas y sin perjuicio que desentrañaré las otras (3), puedo decir a este altura, que todos los delitos (materias) que no tienen término, obedecen a situaciones a las que el siagj no da respuesta, y dependiendo de los criterios de registro de la tramitación, dan origen a esta falta de término.
 52-2018 , uno de los delitos de la acusación se subsumió (incluyó, por decirlo de algún modo) en el otro, por tanto, el delito subsumido no llevó pena. En esa época, el criterio era dar cuenta de todos los delitos de la acusación, y registrar qué pasó con ellos, y para este caso no había cómo registrar y queda sin término. Actualmente simplemente los eliminamos y ya no damos cuenta de los delitos de la acusación versus sentencia, sino solo acusación.
 111-2017 No hay término para un  delito respecto del cual el tribunal se declaró incompetente.
 155-2017 , caso parecido al primero en el efecto, aunque no en lo jurídico. Acá hubo un concurso aparente de leyes penales, es decir una hecho que puede encuadrarse en dos delitos, y se acusó por los dos , se resolvió prefiriendo uno de ellos, el otro queda sin término."

df_fulldata_materia = df_fulldata_materia.progress_apply(transformdata.faltantes_materia, axis=1)

columnas_drop = ['TIPO CAUSA_y', 'MATERIA_y', 'TRIBUNAL_y', 'COD. CORTE_y', 'CORTE_y', 'FECHA INGRESO_y']

df_fulldata_materia.drop(columnas_drop, axis = 'columns', inplace = True)

df_fulldata_materia.rename(columns = {'COD. CORTE_x':'COD. CORTE',
                                      'CORTE_x':'CORTE',
                                      'TRIBUNAL_x':'TRIBUNAL',
                                      'TIPO CAUSA_x':'TIPO CAUSA',
                                      'MATERIA_x':'MATERIA',
                                      'FECHA INGRESO_x':'FECHA INGRESO'
                                     }, inplace = True)

df_fulldata_materia.sample(20)

df_fulldata_materia['AÑO INGRESO'].value_counts()

print(f"Existen un total de {len(df_fulldata_materia)} causas por materias")

De los cuales :

df_fulldata_materia['MOTIVO TERMINO'].value_counts()

Acá analizamos si existe alguna relación entre Total Ingresos por Matería y Total Terminos.

-----------------------------------------------------------------

filtro_oral = df_fulldata_materia[df_fulldata_materia['TRIBUNAL'].str.contains('ORAL')]
filtro_garantia = df_fulldata_materia[df_fulldata_materia['TRIBUNAL'].str.contains('GARANTIA')]

filtro_oral.sample(10)

filtro_garantia.sample(10)

filtro_oral['TRIBUNAL'].value_counts()

filtro_oral['MATERIA'].value_counts()

filtro_oral['MOTIVO TERMINO'].value_counts()

## CARGA DATA A ARCHIVOS FUTHER

# Directorio donde se guardaran archivos feather
df_fulldata_materia.reset_index(inplace = True)
filtro_oral.reset_index(inplace = True)
filtro_garantia.reset_index(inplace = True)
os.makedirs(path_processed, exist_ok = True)  

# Guardamos dataset como archivo feather

df_fulldata_materia.to_feather(f'{path_processed}/consolidated_Materia.feather')
filtro_oral.to_feather(f'{path_processed}/consolidated_JuicioOralesMateria.feather')
filtro_garantia.to_feather(f'{path_processed}/consolidated_CausasGarantiaMateria.feather')
