==> notebooks/6.2-jalvaradoruiz-consolidacion-dataset-rol.ipynb <==
# Creacion Consolidado por ROL

Comenzaremos el analisis de la data consolidando la info en un unico DataSet, para ello vamos a considerar los siguientes puntos:


- DATA:  Se considerará toda las causas que ingresaron al Sistema desde 2015 al 2019, esto ya que se cuenta además con la data de las audiencias en ese período.


- No se consideran causas anteriores a este período.

- Se considerarán Causas sin termino.

import pandas as pd
import numpy as np
import os
import warnings

from tqdm import tqdm
from unicodedata import normalize
from pjud.data import cleandata
from pjud.data import transformdata

warnings.filterwarnings(action="ignore")
tqdm.pandas()

pd.set_option('display.max_columns', 100)

path_processed = "../data/processed/pjud"

#### Paso 1 : Consolidacion de Dataset x rol

df_ingresos_rol = pd.read_feather(f"{path_processed}/processes_IngresosRol.feather")
print(f"{len(df_ingresos_rol)} causas en el dataset de ingresos por Rol")

df_termino_rol = pd.read_feather(f"{path_processed}/processes_TerminosRol.feather")
print(f"Existen : {len(df_termino_rol)} causas en el dataset de termino por rol")

Analizamos la data, uniendo ambos dataset.

df_fulldata_rol = pd.merge(df_ingresos_rol, df_termino_rol, how='outer', on=['COD. TRIBUNAL','RIT'])
print(f"Existen: {len(df_fulldata_rol)} causas en total uniendo ambos dataset, aca existen datos sin registro de ingreso")

df_fulldata_rol['AÑO INGRESO'].value_counts()

### Analisis de la Data Causas por Rol

De estos dos dataset debo considerar:

- Caso 1: el dataset df_causas_terminadas completo
- Caso 2: Debo agregar además las causas que tengan Fecha Ingreso_x pero que aún no terminan.
- Caso 3: Debo analizar las causas que no tienen Fecha Ingreso_x pero si tienen Fecha Ingreso_y y Fecha Ingreso_y, ya que deben ser las causas que no fueron tomadas por la consulta de Ingresos. Acá se debe reemplazar los datos obtenidos en Termino Rol a Ingresos.

Luego de tener estos tres dataset debo concatenarlos

df_fulldata_rol.sample(10)

df_fulldata_rol.columns

columnas_drop = ['index_x', 'index_y', 'MES INGRESO', 'MES TERMINO']

df_fulldata_rol.drop(columnas_drop, axis = 'columns', inplace = True)

df_fulldata_rol = df_fulldata_rol.progress_apply(transformdata.faltantes_rol, axis=1)

columnas_drop = ['TIPO CAUSA_y', 'TRIBUNAL_y', 'COD. CORTE_y', 'CORTE_y', 'FECHA INGRESO_y']

df_fulldata_rol.drop(columnas_drop, axis = 'columns', inplace = True)

df_fulldata_rol.rename(columns = {'COD. CORTE_x':'COD. CORTE',
                                  'CORTE_x':'CORTE',
                                  'TRIBUNAL_x':'TRIBUNAL',
                                  'TIPO CAUSA_x':'TIPO CAUSA',
                                  'MATERIA_x':'MATERIA',
                                  'FECHA INGRESO_x':'FECHA INGRESO'
                                  }, inplace = True)

df_fulldata_rol.sample(10)

df_fulldata_rol['AÑO INGRESO'].value_counts()

df_fulldata_rol['MOTIVO TERMINO'].value_counts()

print(f"Existen: {len(df_fulldata_rol)} causas por rol")

causas_top = df_fulldata_rol[df_fulldata_rol['TRIBUNAL'].str.contains('ORAL')]
causas_garantia = df_fulldata_rol[df_fulldata_rol['TRIBUNAL'].str.contains('GARANTIA')]


causas_top.sample(10)

causas_top.query("RIT == '309-2019' and `COD. TRIBUNAL`== 1323")

causas_garantia.sample(10)

causas_top['MOTIVO TERMINO'].value_counts()

causas_garantia['MOTIVO TERMINO'].value_counts()

ATENCION:
Se detecta que existen RIT 0-AÑO, lo cual no existe en la realidad, estos RIT deberían corresponder a errores del Sistema o ajustes para Metas de Gestión. Para efecto de este analisis deberían ELIMINARSE.

df_rit_cero = df_fulldata_rol[df_fulldata_rol['RIT'].str.startswith("0-")]

df_rit_cero

df_fulldata_rol.drop(df_rit_cero.index, axis=0, inplace=True)

## CARGA DATA A ARCHIVOS FUTHER

# Directorio donde se guardaran archivos feather
df_fulldata_rol.reset_index(inplace = True)
causas_top.reset_index(inplace = True)
causas_garantia.reset_index(inplace = True)
os.makedirs(path_processed, exist_ok = True)  

# Guardamos dataset como archivo feather

df_fulldata_rol.to_feather(f'{path_processed}/consolidated_Rol.feather')
causas_top.to_feather(f'{path_processed}/consolidated_JuicioOralesRol.feather')
causas_garantia.to_feather(f'{path_processed}/consolidated_CausasGarantiaRol.feather')
