==> notebooks/6.3-jalvaradoruiz-consolidacion-dataset-materia-rol.ipynb <==
# Creacion Consolidado por Materia Penal y Rol

Comenzaremos el analisis de la data consolidando la info en un unico DataSet

import pandas as pd
import numpy as np
import os
import warnings

from tqdm import tqdm
from unicodedata import normalize

warnings.filterwarnings(action="ignore")
pd.set_option('display.max_columns', 100)

path_processed = "../data/processed/pjud"

df_rol = pd.read_feather(f"{path_processed}/consolidated_Rol.feather")
print(f"Existen : {len(df_rol)} causas en el dataset por Rol")

df_rol.columns

df_materia = pd.read_feather(f"{path_processed}/consolidated_Materia.feather")
print(f"{len(df_materia)} causas en el dataset por materia")

df_materia.columns

Analizamos la data, uniendo ambos dataset.

df_union = pd.merge(df_rol, df_materia, how='left', on=['COD. CORTE','COD. TRIBUNAL','RIT'], indicator=True)
print(f"Existen: {len(df_union)} causas en total uniendo ambos dataset")

df_union.sample(10)

df_union['AÑO INGRESO_x'].value_counts()

df_union['_merge'].value_counts()

df_sin_materia = df_union.query("_merge == 'left_only'")

Existen Causas sin Materia ... 

df_sin_materia

len(df_sin_materia)

Podemos ver que existen causas que estan con ROL pero no han sido incorporadas en causas por Materia.

¿Que hacer con ello?

A mi parecer se deben desestimar para el analisis, los datos son inherentes por su cantidad.

df_sin_materia['MOTIVO TERMINO_x'].value_counts()

df_sin_materia['TRIBUNAL_x'].value_counts()

df_sin_materia['AÑO INGRESO_x'].value_counts()

Estas no deberian tomarse en cuenta, ya que pueden deberse a causas mal tramitadas.

df_union.columns

df_union.sample(3)

columnas_duplicadas = ['index_x', 'index_y','CORTE_y', 'TRIBUNAL_y',
                       'TIPO CAUSA_y', 'FECHA INGRESO_y',
                       'AÑO INGRESO_y', 'FECHA TERMINO_y',
                       'AÑO TERMINO_y', 'MOTIVO TERMINO_y','DURACION CAUSA_y',
                       'TOTAL TERMINOS_y', '_merge']

df_union.drop(columnas_duplicadas, axis='columns', inplace=True)

df_union.columns

df_union.rename(columns = {'CORTE_x':'CORTE',
                           'TRIBUNAL_x':'TRIBUNAL',
                           'TIPO CAUSA_x':'TIPO CAUSA',
                           'FECHA INGRESO_x':'FECHA INGRESO',
                           'AÑO INGRESO_x':'AÑO INGRESO',
                           'FECHA TERMINO_x':'FECHA TERMINO',
                           'AÑO TERMINO_x':'AÑO TERMINO',
                           'MOTIVO TERMINO_x':'MOTIVO TERMINO',
                           'DURACION CAUSA_x':'DURACION CAUSA',
                           'TOTAL TERMINOS_x':'TOTAL TERMINOS'
                          },inplace = True)

df_union

## CARGA DATA A ARCHIVOS FUTHER

# Directorio donde se guardaran archivos feather
df_union.reset_index(inplace = True)
os.makedirs(path_processed, exist_ok = True)  

# Guardamos dataset como archivo feather

df_union.to_feather(f'{path_processed}/consolidated_Materia_Rol.feather')
