==> notebooks/6.4-jalvaradoruiz-consolidado-causas.ipynb <==
# Creacion Consolidados Causas

Comenzaremos el analisis de la data consolidando la info en un unico DataSet

import pandas as pd
import numpy as np
import os
import warnings

from tqdm import tqdm
from unicodedata import normalize
from pjud.data import transformdata
from pjud.data import cleandata

warnings.filterwarnings(action="ignore")
pd.set_option('display.max_columns', 100)

tqdm.pandas()

path_processed = "../data/processed/pjud"

df_causas = pd.read_feather(f"{path_processed}/consolidated_Materia_Rol.feather")
print(f"{len(df_causas)} causas en el dataset por materia y rol")

df_causas.columns

df_causas['TRIBUNAL-RIT'] = df_causas['COD. TRIBUNAL'].map(str) + "-" + df_causas['RIT'].map(str)


len(df_consolidado)

df_consolidado.columns

df_consolidado._merge.value_counts()

años = range(2015,2020)
consolidado_audiencias = []
#df_consolidado_audiencias = df_consolidado.set_index("FECHA AUDIENCIA")

for año in años:
    perido_inicial = '01-01-' + str(año)
    periodo_final = '31-12-' + str(año)
    mask_periodo = f"`FECHA AUDIENCIA` >='{perido_inicial}' and `FECHA AUDIENCIA` <= '{periodo_final}'"
    total_año = len(df_consolidado.query(mask_periodo))
    print(f"{mask_periodo} -> {total_año}")
    consolidado_audiencias.append(total_año)

consolidado_audiencias

audiencias_sin_causas = df_consolidado.query("_merge == 'right_only'")
len(audiencias_sin_causas)

audiencias_sin_causas.sample(8)

Imputo algunos datos que si dispongo del dataset.



audiencias_sin_causas['COD. CORTE_x'] = audiencias_sin_causas['COD. CORTE_y']
audiencias_sin_causas['CORTE_x'] = audiencias_sin_causas['CORTE_y']
audiencias_sin_causas['TRIBUNAL_x'] = audiencias_sin_causas['TRIBUNAL_y']
audiencias_sin_causas['TIPO CAUSA_x'] = audiencias_sin_causas['TIPO CAUSA_y']

audiencias_sin_causas['AÑO INGRESO'] = audiencias_sin_causas['RIT'].progress_apply(cleandata.obtiene_año)




audiencias_sin_causas

Elimino de consolidado el original de audiencias_sin_causas

df_consolidado.drop(audiencias_sin_causas.index, inplace=True)

Compruebo verificando el _merge --> debe ser 0 en right_only

df_consolidado['_merge'].value_counts()

df_consolidado = pd.concat([df_consolidado,audiencias_sin_causas])

df_consolidado['_merge'].value_counts()

df_consolidado.query("_merge=='right_only'")

sin_audiencias = df_consolidado[df_consolidado['TIPO DE AUDIENCIA'].isnull()]

sin_audiencias['AÑO INGRESO'].value_counts()

sin_audiencias.sample(10)

len(sin_audiencias)

sin_audiencias['MOTIVO TERMINO'].value_counts()

sin_audiencias['TRIBUNAL_x'].value_counts()

sin_audiencias['TRIBUNAL_x'].value_counts().sum() , sin_audiencias['TRIBUNAL_y'].value_counts().sum()

mask_tribunal = sin_audiencias['TRIBUNAL_x']!=sin_audiencias['TRIBUNAL_y']
sin_audiencias[mask_tribunal]


mask_materi_null = df_consolidado['MATERIA'].isnull()
sin_materia = df_consolidado[mask_materi_null]

len(sin_materia)

sin_materia

sin_materia["_merge"].value_counts()

## Carga Data relacionada a Tipologia de delitos con CausaS

path = "../data/processed/delitos"
df_tipologia = pd.read_feather(f"{path}/clean_Delitos.feather")
print(f"{len(df_tipologia)} delitos utilizados en PJUD")

df_tipologia.columns

df_causa_tipologia = pd.merge(df_causas,df_tipologia, how='left', on=['COD. MATERIA'])

df_causa_tipologia.sample(10)

columnas_duplicadas = ['index_x', 'MATERIA_x','index_y']

df_causa_tipologia.drop(columnas_duplicadas, axis='columns', inplace=True)



df_causa_tipologia.rename(columns = {'MATERIA_y':'MATERIA'
                                    }, inplace=True
                         )

df_causa_tipologia.columns

## Carga Data relacionada a Poblacion

df_poblacion = pd.read_feather(f"{path_processed}/processes_DataConsolidada_Poblacion_Jurisdiccion.feather")
print(f"{len(df_poblacion)} registros")

df_poblacion

df_poblacion.query("CORTE == 'C.A. DE SAN MIGUEL'")

df_poblacion.query("CORTE == 'C.A. DE SANTIAGO'")

en_tribunales = df_poblacion['TRIBUNAL'].unique()

en_union = df_union_tipologia['TRIBUNAL'].unique()

for tribunal in df_union_tipologia['TRIBUNAL'].unique():
    if not tribunal in en_tribunales:
        print(tribunal)

df_poblacion['CORTE'].unique()

df_union_tipologia['CORTE'].value_counts()

mask_cortenull = df_union_tipologia['CORTE'].isnull()
len(df_union_tipologia[mask_cortenull])

df_fulldatacausa = pd.merge(df_causa_tipologia, df_poblacion, how='left', on=['CORTE','TRIBUNAL'])

len(df_fulldatacausa)

sin_region = df_fulldatacausa[df_fulldatacausa['REGION'].isnull()]

sin_region['CORTE'].value_counts()

df_fulldatacausa

mask_sincortes = df_fulldata['COD. CORTE_x'].isnull()
sin_cortes = df_fulldata[mask_sincortes]
sin_cortes

df_fulldata['MATERIA_x'].notnull().sum(), df_fulldata['MATERIA_y'].notnull().sum()

df_fulldatacausa.columns

VERIFICAR FILAS A ELIMINAR ... CAMBIARON ....

columnas_duplicadas = ['index']

df_fulldatacausa.drop(columnas_duplicadas, axis='columns', inplace=True)


#df_consolidado.drop(['MATERIA_x','FECHA TERMINO_x','AÑO TERMINO_x','TIPO CAUSA_y'], axis='columns', inplace=True)

df_fulldatacausa.sample(5)

Reordenando Nombres de las columnas ...

df_fulldatacausa.rename(columns = {'COD. CORTE':'cod_corte',
                                 'COD. TRIBUNAL':'cod_tribunal',
                                 'RIT':'rit',
                                 'COD. MATERIA':'cod_materia',
                                 'TOTAL INGRESOS POR MATERIAS':'total_ingresos_materia',
                                 'FECHA INGRESO':'fecha_ingreso',
                                 'AÑO INGRESO':'año_ingreso',
                                 'FECHA TERMINO':'fecha_termino',
                                 'DURACION CAUSA':'duracion_causa',
                                 'MOTIVO TERMINO':'motivo_termino',
                                 'AÑO TERMINO':'año_termino',
                                 'TOTAL TERMINOS':'total_terminos',
                                 'CORTE':'corte',
                                 'TRIBUNAL':'tribunal',
                                 'TIPO CAUSA':'tipo_causa',
                                 'TRIBUNAL-RIT':'tribunal_rit',
                                 'MATERIA':'materia',
                                 'TIPOLOGIA MATERIA':'tipologia_materia',
                                 'VIGENCIA MATERIA':'vigencia_materia',
                                 'REGION':'region',
                                 'POBLACION':'poblacion',
                                 'HOMBRES':'hombres',
                                 'MUJERES':'mujeres',
                                 'URBANO':'urbano',
                                 'RURAL':'rural',
                                 'COMUNAS':'comunas',
                                 'JUECES':'dotacion_jueces',
                                 'ASIENTO':'asiento',
                                 'TIPO JUZGADO':'tipo_juzgado'
                                 },inplace = True)

df_fulldatacausa.columns

df_fulldatacausa = df_fulldatacausa[['region','cod_corte','corte','tribunal_rit','cod_tribunal','rit','tribunal','tipo_juzgado','dotacion_jueces','tipo_causa','fecha_ingreso','año_ingreso','cod_materia','materia',
                                     'tipologia_materia','vigencia_materia','total_ingresos_materia','total_terminos','fecha_termino','año_termino','duracion_causa','motivo_termino','asiento','comunas','poblacion',
                                     'hombres','mujeres','urbano','rural']]

df_fulldatacausa.sample(5)

# Directorio donde se guardaran archivos feather
df_fulldatacausa.reset_index(inplace = True)
os.makedirs(path_processed, exist_ok = True)  

# Guardamos dataset como archivo feather

df_fulldatacausa.to_feather(f'{path_processed}/consolidated_FULLDATA_CAUSAS.feather')
